Move command into proper package
